+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- fix: notebook dependency failure (#78)
+
 ## [0.6.0] - 2021-12-31
 
 ### Fixed